added retry testing util
